Update MAGIC_Act.py
Improved resistance to reusing same temporary assignments.
Peaks with types containing L and/or V are colored purple if a pari cannot be formed  using the HMBC-HMQC
Check NOESY now examines the symmetry of the NOESY cross peaks observed for geminal pairs, including overlapped peaks, and colors them green if >= 85% of cross peaks are shared, gold if > 60% of cross peaks are shared, and magenta if less than 60% of cross peaks are shared. The score is weighted by signal to noise, so that missing low intensity cross peaks are penalized less. 
A Generate FLYA option was added to produce the files needed to run Methyl FLYA. 
Other scripting refinements were made that did not change any performance features